% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armagarch.R
\name{armagarchf}
\alias{armagarchf}
\title{ARMA(1, 1)-GARCH(1, 1) forecasting (with simulation)}
\usage{
armagarchf(
  y,
  h = 5,
  level = 95,
  B = NULL,
  cl = 1L,
  dist = c("student", "gaussian"),
  seed = 123,
  ...
)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{number of periods for forecasting}

\item{level}{confidence level for prediction intervals}

\item{B}{number of simulations for \code{arima.sim}}

\item{cl}{an integer; the number of clusters for parallel execution}

\item{dist}{distribution of innovations ("student" or "gaussian")}

\item{seed}{reproducibility seed}

\item{...}{Additional arguments passed to \code{fGarch::garchFit}}
}
\value{
An object of class "forecast"; a list containing the following elements:

\item{model}{A list containing information about the fitted model}
\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{x}{The original time series}
\item{sims}{Simulations of ARMA(1, 1)-GARCH(1, 1)}
}
\description{
ARMA(1, 1)-GARCH(1, 1) forecasting (with simulation)
}
\examples{

y <- datasets::EuStockMarkets[ , "DAX"]

# require(forecast)
# z <- ahead::armagarchf(y=y, h=200)
# plot(z)

}
\author{
T. Moudiki
}
