% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basicf}
\alias{basicf}
\title{Basic forecasting (mean, median, random walk)}
\usage{
basicf(
  y,
  h = 5,
  level = 95,
  method = c("mean", "median", "rw"),
  type_pi = c("gaussian", "bootstrap", "blockbootstrap", "movingblockbootstrap"),
  block_length = NULL,
  seed = 1,
  B = 100,
  show_progress = TRUE
)
}
\arguments{
\item{y}{A multivariate time series of class \code{ts} or a matrix}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{method}{forecasting method, either "mean", "median", or random walk ("rw")}

\item{type_pi}{type of prediction interval currently, "gaussian", "bootstrap",
"blockbootstrap" or "movingblockbootstrap"}

\item{block_length}{length of block for (circular) "blockbootstrap" or
"movingblockbootstrap"}

\item{seed}{reproducibility seed for \code{type_pi == 'bootstrap'}}

\item{B}{Number of bootstrap replications for \code{type_pi == 'bootstrap'}}

\item{show_progress}{A boolean; show progress bar for bootstrapping? Default is TRUE.}
}
\value{
An object of class "mtsforecast"; a list containing the following elements:

\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{sims}{Model simulations for bootstrapping (basic, or block)}
\item{x}{The original time series}
\item{residuals}{Residuals from the fitted model}
\item{coefficients}{Regression coefficients for \code{type_pi == 'gaussian'} for now}
}
\description{
Basic forecasting functions for multivariate time series
}
\examples{

\dontrun{

require(fpp2)

res <- ahead::basicf(fpp2::insurance, h=10)
par(mfrow=c(1, 2))
plot(res, "TV.advert")
plot(res, "Quotes")


res <- ahead::basicf(fpp2::insurance, method="rw", h=10)
par(mfrow=c(1, 2))
plot(res, "TV.advert")
plot(res, "Quotes")


# block bootstrap
res3 <- ahead::basicf(fpp2::insurance, h=10, type_pi = "bootstrap", B=10)
res5 <- ahead::basicf(fpp2::insurance, h=10, type_pi = "blockbootstrap", B=10,
                      block_length = 4)

print(res3$sims[[2]])
print(res5$sims[[2]])

par(mfrow=c(2, 2))
plot(res3, "Quotes")
plot(res3, "TV.advert")
plot(res5, "Quotes")
plot(res5, "TV.advert")


# moving block bootstrap
res6 <- ahead::basicf(fpp2::insurance, h=10,
                      type_pi = "movingblockbootstrap", B=10,
                      block_length = 4, method = "rw")

par(mfrow=c(1, 2))
plot(res6, "Quotes")
plot(res6, "TV.advert")

}

}
