% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_GLMNET.R
\name{comb_GLMNET}
\alias{comb_GLMNET}
\title{GLMNET Regression Forecast Combination}
\usage{
comb_GLMNET(x, custom_error = NULL)
}
\arguments{
\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}

\item{custom_error}{Custom error function}
}
\value{
Returns an object of class \code{ForecastComb::foreccomb_res} with the following components:
\item{Method}{Returns the best-fit forecast combination method.}
\item{Models}{Returns the individual input models that were used for the forecast combinations.}
\item{Weights}{Returns the combination weights obtained by applying the combination method to the training set.}
\item{Intercept}{Returns the intercept of the linear regression.}
\item{Fitted}{Returns the fitted values of the combination method for the training set.}
\item{Accuracy_Train}{Returns range of summary measures of the forecast accuracy for the training set.}
\item{Forecasts_Test}{Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.}
\item{Accuracy_Test}{Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.}
\item{Input_Data}{Returns the data forwarded to the method.}
}
\description{
Computes forecast combination weights using GLMNET Regression (OLS) regression.
}
\details{
The function integrates the GLMNET Regression forecast combination implementation of the
\emph{ForecastCombinations} package into ForecastComb.

The results are stored in an object of class 'ForecastComb::foreccomb_res', for which separate plot and summary functions are provided.
}
\examples{

\dontrun{

library(ForecastComb)

data(electricity)

print(head(electricity))

forecasting_methods <- colnames(electricity)[1:5]

train_obs <- electricity[1:84, "Actual"]
train_pred <- electricity[1:84, forecasting_methods]
test_obs <- electricity[85:123, "Actual"]
test_pred <- electricity[85:123, forecasting_methods]
data <- ForecastComb::foreccomb(train_obs, train_pred, test_obs, test_pred)

obj <- ahead::comb_GLMNET(data))

}

}
\keyword{models}
