% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_attention.R
\name{computeattention}
\alias{computeattention}
\title{Compute global attention weights and context vectors for time series}
\usage{
computeattention(series)
}
\arguments{
\item{series}{Numeric vector containing the time series of length n}
}
\value{
List containing:
\item{attention_weights}{n × n matrix where entry (i,j) represents the attention
weight of time j on time i. Only entries j <= i are non-zero (causal attention).}
\item{context_vectors}{Vector of length n where each entry i is the weighted sum
of all values up to time i, using the attention weights.}
}
\description{
Compute global attention weights and context vectors for time series
}
\examples{
# For a series of length 5
series <- c(1, 2, 3, 4, 5)
result <- compute_attention(series)

# attention_weights will be 5x5 matrix
# context_vectors will be length 5
dim(result$attention_weights)  # [1] 5 5
length(result$context_vectors) # [1] 5
}
