% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts2.R
\name{create_2level_hts}
\alias{create_2level_hts}
\title{Create simple 2-level hierarchical time series}
\usage{
create_2level_hts(bts)
}
\arguments{
\item{bts}{Bottom-level time series matrix (T x m)}
}
\value{
List with total and bottom series
}
\description{
Create simple 2-level hierarchical time series
}
