% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createtrendseason}
\alias{createtrendseason}
\title{Create trend and seasonality features for univariate time series}
\usage{
createtrendseason(y)
}
\arguments{
\item{y}{a univariate time series}
}
\value{
a vector or matrix of features
}
\description{
Create trend and seasonality features for univariate time series
}
\examples{

y <- ts(rnorm(100), start = 1, frequency = 12)
createtrendseason(y)

createtrendseason(USAccDeaths)

}
