% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrm_sensi.R
\name{dynrmf_sensi}
\alias{dynrmf_sensi}
\title{Compute First-Order Sensitivity Effects for Dynamic Regression Forecasts}
\usage{
dynrmf_sensi(y, xreg, h = NULL, level = 99, zero = 1e-04, ...)
}
\arguments{
\item{y}{A univariate time series object to forecast.}

\item{xreg}{A time series of external regressors (same length as \code{y}).}

\item{h}{Forecast horizon (length of test set). If NULL, uses length of xreg.}

\item{level}{Confidence level for \code{dynrmf} (default 99).}

\item{zero}{Small positive constant to avoid division by zero (default 1e-4).}

\item{...}{Additional parameters to be passed to \code{\link{ahead::dynrmf}}}
}
\value{
A list containing:
\describe{
\item{effects_mean}{Sensitivity effects on the mean forecast.}
\item{forecast_central}{Forecast without xreg.}
\item{forecast_with_xreg}{Forecast with xreg.}
\item{forecast_plus}{Forecast with positive perturbation on xreg.}
\item{forecast_minus}{Forecast with negative perturbation on xreg.}
\item{h}{Forecast horizon used.}
\item{xreg_test}{Test set of xreg used in forecasting.}
\item{hh}{Perturbation applied to xreg.}
\item{split}{List with training and testing sets of y.}
}
}
\description{
This function computes the sensitivity of a forecasted time series (\code{y})
to small perturbations in an external regressor (\code{xreg}) using the \code{dynrmf} function.
}
