% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_benchmarks.R
\name{fitforecast}
\alias{fitforecast}
\title{Fit anf forecast for benchmarking purposes}
\usage{
fitforecast(
  y,
  h = NULL,
  pct_train = 0.9,
  pct_calibration = 0.5,
  method = c("arima", "ets", "tbats", "tslm", "dynrmf", "ridge2f", "naive", "snaive",
    "thetaf"),
  level = 95,
  B = 1000L,
  seed = 17223L,
  graph = TRUE,
  conformalize = FALSE,
  type_calibration = c("splitconformal", "cv1", "loocv"),
  gap = 3L,
  agg = c("mean", "median"),
  vol = c("constant", "garch"),
  type_sim = c("kde", "surrogates", "bootstrap"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fit anf forecast for benchmarking purposes
}
\examples{

plot(ahead::fitforecast(AirPassengers))


}
