% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genforecast.R
\name{genericforecast}
\alias{genericforecast}
\title{Generic Forecasting Function (Unified interface)}
\usage{
genericforecast(FUN, y, h, level = 95, ...)
}
\arguments{
\item{FUN}{A forecasting function.}

\item{y}{A time series (\code{ts} object or vector).}

\item{h}{Forecasting horizon.}

\item{level}{The confidence level.}

\item{...}{Additional arguments to be passed to the forecasting function.}
}
\value{
An object of class \code{forecast}.
}
\description{
This function allows to call any function "of class \code{forecast}" in a unified way.
}
\examples{

y <- fdeaths 
h <- 25L
plot(ahead::genericforecast(FUN=forecast::thetaf, y, h))
plot(ahead::genericforecast(FUN=ahead::dynrmf, y, h))
# plot(ahead::genericforecast(FUN=forecast::tbats, y=y, h=h, use.box.cox = TRUE, use.trend=FALSE))
# plot(ahead::genericforecast(FUN=forecast::auto.arima, y, h)) 

}
