% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_benchmarks.R
\name{get_error}
\alias{get_error}
\title{Get error metrics}
\usage{
get_error(obj, actual, level = 95)
}
\arguments{
\item{obj}{A prediction object containing mean predictions and intervals.}

\item{actual}{A numeric vector of actual values.}

\item{level}{The confidence level for the prediction intervals.}
}
\value{
A numeric vector containing the error metrics.
}
\description{
This function calculates various error metrics for a given prediction object.
}
\examples{

obj <- list(mean = 10, lower = 8, upper = 12)
actual <- 11
get_error(obj, actual)

}
