% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getreturns.R
\name{getreturns}
\alias{getreturns}
\title{Calculate returns or log-returns for multivariate time series}
\usage{
getreturns(x, type = c("basic", "log"))
}
\arguments{
\item{x}{Multivariate time series}

\item{type}{Type of return: basic return ("basic") or log-return ("log")}
}
\value{
The returns
}
\description{
Calculate returns or log-returns for multivariate time series
}
\examples{

returns <- getreturns(EuStockMarkets)
log_returns <- getreturns(EuStockMarkets,
                          type = "log")

par(mfrow=c(1, 3))
matplot(EuStockMarkets, type = 'l', main = "Closing Prices of \n European stocks (1991-1998)",
xlab = "time")
matplot(returns, type = 'l', main = "Returns", xlab = "time")
matplot(log_returns, type = 'l', main = "Log-returns", xlab = "time")

}
