% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsimulations.R
\name{getsimulations}
\alias{getsimulations}
\title{Obtain simulations (when relevant) from a selected time series}
\usage{
getsimulations(obj, selected_series, transpose = FALSE)
}
\arguments{
\item{obj}{result from ridge2f (multivariate time series forecast with simulations)}

\item{selected_series}{name of the time series selected}

\item{transpose}{return a transposed time series}
}
\description{
Obtain simulations (when relevant) from a selected time series
}
\examples{

require(fpp)

obj <- ahead::ridge2f(fpp::insurance, h = 7,
                      type_pi = "bootstrap", B = 5)
print(getsimulations(obj, selected_series = "TV.advert"))
print(getsimulations(obj, selected_series = "Quotes"))
print(getsimulations(obj, selected_series = "TV.advert", transpose = TRUE))
print(getsimulations(obj, selected_series = "Quotes", transpose = TRUE))

}
