% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{loessf}
\alias{loessf}
\title{Loess forecasting}
\usage{
loessf(
  y,
  h = 5,
  level = 95,
  span = 0.75,
  degree = 2,
  b = NULL,
  B = 250,
  seed = 123,
  type_boot = c("pkgforecast", "pkgahead")
)
}
\arguments{
\item{seed}{}
}
\description{
Loess forecasting
}
\examples{

plot(loessf(Nile, h=20, level=95))

}
