% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocvridge2.R
\name{loocvridge2f}
\alias{loocvridge2f}
\title{LOOCV for Ridge2 model}
\usage{
loocvridge2f(
  y,
  xreg = NULL,
  h = 5,
  level = 95,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1,
  dropout = 0,
  type_forecast = c("recursive", "direct"),
  type_pi = c("gaussian", "bootstrap", "blockbootstrap", "movingblockbootstrap",
    "rvinecopula", "splitconformal"),
  block_length = NULL,
  margins = c("gaussian", "empirical", "student"),
  seed = 1,
  B = 100L,
  type_aggregation = c("mean", "median"),
  centers = NULL,
  type_clustering = c("kmeans", "hclust"),
  ym = NULL,
  cl = 1L,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{y}{A multivariate time series of class \code{ts} (preferred) or a \code{matrix}}

\item{xreg}{External regressors. A data.frame (preferred) or a \code{matrix}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags}

\item{nb_hidden}{Number of nodes in hidden layer}

\item{nodes_sim}{Type of simulation for nodes in the hidden layer}

\item{activ}{Activation function}

\item{a}{Hyperparameter for activation function "leakyrelu", "elu"}

\item{lambda_1}{Regularization parameter for original predictors}

\item{lambda_2}{Regularization parameter for transformed predictors}

\item{dropout}{dropout regularization parameter (dropping nodes in hidden layer)}

\item{type_forecast}{Recursive or direct forecast}

\item{type_pi}{Type of prediction interval currently "gaussian", "bootstrap",
"blockbootstrap", "movingblockbootstrap", "splitconformal" (very experimental right now),
"rvinecopula" (with Gaussian margins for now, Student-t coming soon)}

\item{block_length}{Length of block for circular or moving block bootstrap}

\item{margins}{Distribution of margins: "gaussian", "empirical", "student" (postponed or
never) for \code{type_pi == "rvinecopula"}}

\item{seed}{Reproducibility seed for random stuff}

\item{B}{Number of bootstrap replications or number of simulations (yes, 'B' is unfortunate)}

\item{type_aggregation}{Type of aggregation, ONLY for bootstrapping; either "mean" or "median"}

\item{centers}{Number of clusters for \code{type_clustering}}

\item{type_clustering}{"kmeans" (K-Means clustering) or "hclust" (Hierarchical clustering)}

\item{ym}{Univariate time series (\code{stats::ts}) of yield to maturities with
\code{frequency = frequency(y)} and \code{start = tsp(y)[2] + 1 / frequency(y)}.
Default is \code{NULL}.}

\item{cl}{An integer; the number of clusters for parallel execution, for bootstrap}

\item{show_progress}{A boolean; show progress bar for bootstrapping? Default is TRUE.}

\item{...}{Additional parameters to be passed to \code{\link{kmeans}} or \code{\link{hclust}}}
}
\value{
An object of class "mtsforecast"; a list containing the following elements:

\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{sims}{Model simulations for bootstrapping (basic, or block)}
\item{x}{The original time series}
\item{residuals}{Residuals from the fitted model}
\item{coefficients}{Regression coefficients for \code{type_pi == 'gaussian'} for now}
}
\description{
LOOCV for Random Vector functional link network model with 2 regularization parameters
}
\examples{

require(fpp)

print(ahead::loocvridge2f(fpp::insurance))
print(ahead::loocvridge2f(fpp::usconsumption))

#foo <- function(xx) ahead::loocvridge2f(fpp::insurance, lambda_1=10^xx[1], lambda_2=10^xx[2])
#(opt <- stats::nlminb(objective=foo, lower=c(-10,-10), upper=c(10,10), start=c(0, 0)))
#print(ahead::loocvridge2f(fpp::insurance, lambda_1=10^opt$par[1], lambda_2=10^opt$par[2]))

}
\references{
Moudiki, T., Planchet, F., & Cousin, A. (2018).
Multiple time series forecasting using quasi-randomized
functional link neural networks. Risks, 6(1), 22. \cr
}
\author{
T. Moudiki
}
