% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml_forecast}
\alias{ml_forecast}
\title{Forecasting using Machine Leaning models}
\usage{
ml_forecast(
  y,
  h,
  level = 95,
  lags = 1,
  fit_func = ahead::ridge,
  predict_func = predict,
  xreg = NULL,
  coeffs = NULL,
  ...
)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags of the input time series considered in the regression}

\item{fit_func}{Fitting function (Statistical/ML model). Default is Ridge regression.}

\item{predict_func}{Prediction function (Statistical/ML model)}

\item{xreg}{External regressor variable}

\item{coeffs}{Coefficients of the fitted model. If provided, a linear combination with the coefficients is used to compute the prediction.}

\item{...}{additional parameters passed to the fitting function \code{fit_func}}
}
\value{
An object of class 'forecast'
}
\description{
Forecasting using Machine Leaning models
}
\examples{

\dontrun{
plot(ahead::ml_forecast(AirPassengers, h=20L))

plot(ahead::ml_forecast(AirPassengers, h=25L, lags=20L, fit_func=glmnet::cv.glmnet)) 

res <- ahead::ml_forecast(USAccDeaths, h=15L, lags=15L)
plot(res)

res <- ahead::ml_forecast(USAccDeaths, fit_func = glmnet::cv.glmnet, h=15L, lags=15L) 
plot(res)

res <- ahead::ml_forecast(USAccDeaths, fit_func = e1071::svm, h=15L, lags=15L) 
plot(res)

res <- ahead::ml_forecast(mdeaths, h=15L, lags=15L)
plot(res)

res <- ahead::ml_forecast(fdeaths, fit_func = glmnet::cv.glmnet, h=15L, lags=25L) 
plot(res)

res <- ahead::ml_forecast(fdeaths, fit_func = randomForest::randomForest, h=15L, lags=15L) 
plot(res)
}

}
