% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{mlf}
\alias{mlf}
\title{Conformalized Forecasting using Machine Leaning models}
\usage{
mlf(
  y,
  h = 5,
  level = 95,
  lags = 15L,
  fit_func = ahead::ridge,
  predict_func = predict,
  stack = FALSE,
  stacking_models = NULL,
  coeffs = NULL,
  type_pi = c("surrogate", "bootstrap", "kde"),
  B = 250L,
  agg = c("mean", "median"),
  seed = 123,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags of the input time series considered in the regression}

\item{fit_func}{Fitting function (Statistical/ML model). Default is Ridge regression.}

\item{predict_func}{Prediction function (Statistical/ML model)}

\item{stack}{Boolean, use stacking regression or not}

\item{stacking_models}{A list of \code{fit_func}s and \code{predict_func}s for and ensemble of stacked models (you should set \code{stack=TRUE})}

\item{coeffs}{Coefficients of the fitted model. If provided, a linear combination with the coefficients is used to compute the prediction.}

\item{type_pi}{Type of prediction interval}

\item{B}{Number of bootstrap replications or number of simulations}

\item{agg}{"mean" or "median" (aggregation method)}

\item{seed}{reproducibility seed}

\item{show_progress}{show progress bar for stacking, if \code{stacking_models} if not \code{NULL}}

\item{...}{additional parameters passed to the fitting function \code{fit_func}}
}
\value{
An object of class 'forecast'
}
\description{
Conformalized Forecasting using Machine Leaning models
}
