% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mtsforecast}
\alias{plot.mtsforecast}
\title{Plot multivariate time series forecast}
\usage{
\method{plot}{mtsforecast}(x, selected_series, ...)
}
\arguments{
\item{x}{result from varf or ridge2f (multivariate time series forecast)}

\item{selected_series}{name of the time series selected for plotting}

\item{...}{additional parameters to be passed to \code{plot}}
}
\description{
Plot multivariate time series forecast
}
\examples{

require(fpp)

fit_obj_VAR <- ahead::varf(fpp::insurance, lags = 2,
h = 10, level = 95)

fit_obj_ridge2 <- ahead::ridge2f(fpp::insurance, lags = 2,
h = 10, level = 95)

par(mfrow=c(2, 2))
plot(fit_obj_VAR, "Quotes")
plot(fit_obj_VAR, "TV.advert")
plot(fit_obj_ridge2, "Quotes")
plot(fit_obj_ridge2, "TV.advert")

}
