% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pairs}
\alias{plot.pairs}
\title{Plot bivariate distributions}
\usage{
\method{plot}{pairs}(x, y = NULL)
}
\arguments{
\item{x}{bivariate matrix or time series}

\item{y}{additional series (default is \code{NULL})}
}
\description{
Plot bivariate distributions
}
\examples{

obj <- ahead::ridge2f(fpp::insurance, h = 10, type_pi = "blockbootstrap",
block_length=5, B = 10)

\dontrun{
plot(obj$residuals, type_graph = "pairs")
}


}
