% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthsimulation.R
\name{plot.synthetic_ts}
\alias{plot.synthetic_ts}
\title{Plot method for synthetic_ts objects}
\usage{
\method{plot}{synthetic_ts}(x, type = "series", which_sim = 1, series_index = 1, ...)
}
\arguments{
\item{x}{synthetic_ts object}

\item{type}{Type of plot: "series", "residuals", "acf", "density", "qq"}

\item{which_sim}{Which simulation to plot (for type = "series")}

\item{series_index}{Which series to plot (for multivariate data)}

\item{...}{Additional arguments passed to plot functions}
}
\description{
Plot method for synthetic_ts objects
}
