% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{predict.ridge}
\alias{predict.ridge}
\title{Predict from Ridge regression}
\usage{
\method{predict}{ridge}(object, newx)
}
\arguments{
\item{object}{object fitted by function `ridge`}

\item{newx}{new examples}
}
\value{
predicted values for \code{newx}
}
\description{
Predict from Ridge regression
}
\examples{

set.seed(123)

n <- 100 ; p <- 2
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

fit_obj <- ahead::ridge(X, y)

n_test <- 10

predict(fit_obj, newx=matrix(rnorm(n_test * p), n_test, p))


}
