% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretfitpredict.R
\name{predict_func}
\alias{predict_func}
\title{Predict univariate time series using caret ML model(for use with \code{dynrmf})}
\usage{
predict_func(obj, newx)
}
\arguments{
\item{obj}{Fitted object}

\item{newx}{Test set data}
}
\description{
Predict univariate time series using caret ML model(for use with \code{dynrmf})
}
