% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch11.R
\name{print.garch11forecast}
\alias{print.garch11forecast}
\title{Print Method for GARCH(1,1) Forecast Objects}
\usage{
\method{print}{garch11forecast}(x, ...)
}
\arguments{
\item{x}{An object of class "garch11forecast" from \code{garch11f}}

\item{...}{Additional arguments passed to print method}
}
\value{
Invisibly returns the input object
}
\description{
Prints a summary of GARCH(1,1) forecasts in a user-friendly format.
}
\seealso{
\code{\link{garch11f}}
}
