% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthsimulation.R
\name{print.summary.synthetic_ts}
\alias{print.summary.synthetic_ts}
\title{Print method for summary.synthetic_ts objects}
\usage{
\method{print}{summary.synthetic_ts}(x, ...)
}
\arguments{
\item{x}{summary.synthetic_ts object}

\item{...}{Additional arguments}
}
\description{
Print method for summary.synthetic_ts objects
}
