% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge2.R
\name{ridge2f}
\alias{ridge2f}
\title{Ridge2 model}
\usage{
ridge2f(
  y,
  h = 5,
  level = 95,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1,
  type_forecast = c("recursive", "direct"),
  type_pi = c("gaussian", "bootstrap"),
  seed = 1,
  B = 100,
  cl = NULL
)
}
\arguments{
\item{y}{A multivariate time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags}

\item{nb_hidden}{Number of nodes in hidden layer}

\item{nodes_sim}{Type of simulation for nodes in the hidden layer}

\item{activ}{Activation function}

\item{a}{hyperparameter for activation function "leakyrelu", "elu"}

\item{lambda_1}{Regularization parameter for original predictors}

\item{lambda_2}{Regularization parameter for transformed predictors}

\item{type_forecast}{Recursive or direct forecast}

\item{type_pi}{currently "gaussian" or "bootstrap"}

\item{seed}{reproducibility seed for \code{type_pi == 'bootstrap'}}

\item{B}{Number of bootstrap replications for \code{type_pi == 'bootstrap'}}

\item{cl}{an integer; the number of clusters for parallel execution, for \code{type_pi == 'bootstrap'}}
}
\value{
An object of class "mtsforecast"; a list containing the following elements:

\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{x}{The original time series}
\item{residuals}{Residuals from the fitted model}
\item{coefficients}{Regression coefficients for \code{type_pi == 'gaussian'} for now}
\item{sims}{Model simulations for \code{type_pi == bootstrap}}
}
\description{
Random Vector functional link network model with 2 regularization parameters
}
\examples{

require(fpp)

print(ahead::ridge2f(fpp::insurance)$mean)
print(ahead::ridge2f(fpp::usconsumption)$lower)

res <- ahead::ridge2f(fpp::insurance, h=10, lags=2)
par(mfrow=c(1, 2))
plot(res, "Quotes")
plot(res, "TV.advert")

res <- ahead::ridge2f(fpp::usconsumption, h=20, lags=2,
lambda_2=1)
par(mfrow=c(1, 2))
plot(res, "income")
plot(res, "consumption")

}
\references{
Moudiki, T., Planchet, F., & Cousin, A. (2018).
Multiple time series forecasting using quasi-randomized
functional link neural networks. Risks, 6(1), 22. \cr
}
\author{
T. Moudiki
}
