% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridgef}
\alias{ridgef}
\title{Forecasts from ridge regression models}
\usage{
ridgef(
  y,
  h = 5,
  level = 95,
  lags = 1,
  nb_hidden = 5,
  fit_method = c("ridge", "mgaussian"),
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  a = 0.01,
  lambda = 0.1,
  alpha = 0.5,
  seed = 1,
  type_pi = "none",
  type_forecast = c("recursive", "direct")
)
}
\arguments{
\item{y}{A multivariate time series of class \code{ts} (preferred) or a \code{matrix}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags to include in the model}

\item{nb_hidden}{Number of hidden units in the neural network}

\item{fit_method}{Method to fit the model. Either \code{ridge} for elastic net or \code{mgaussian} for multivariate Gaussian elastic net}

\item{nodes_sim}{Method to generate the nodes for the simulation of the multivariate . Either \code{sobol}, \code{halton} or \code{unif}}

\item{activ}{Activation function for the hidden layer. Either \code{relu}, \code{sigmoid}, \code{tanh}, \code{leakyrelu}, \code{elu} or \code{linear}}

\item{hidden_layer_bias}{Whether to include a bias term in the hidden layer}

\item{a}{Hyperparameter for activation function "leakyrelu", "elu"}

\item{lambda}{Hyperparameter for elastic net (regularization parameter)}

\item{alpha}{Hyperparameter for elastic net (compromise between ridge and lasso)}

\item{seed}{Seed for reproducibility}

\item{type_forecast}{Type of forecast. Either \code{recursive} or \code{direct}}

\item{type_ci}{Type of confidence interval. Either \code{none}, \code{parametric} or \code{nonparametric}}
}
\value{
an object of class \code{mforecast}; a list containing the following elements:
\item{mean}{Point forecasts for the time series} \item{resid}{Residuals from the fitted model}
\item{model}{A list containing information about the fitted model}
\item{method}{The name of the forecasting method as a character string}
\item{x}{The original time series}
}
\description{
Forecasts from ridge regression models
}
