% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts2.R
\name{simple_forecast}
\alias{simple_forecast}
\title{Simple ETS forecast function}
\usage{
simple_forecast(y, h)
}
\arguments{
\item{y}{Time series vector}

\item{h}{Forecast horizon}
}
\value{
Forecast vector
}
\description{
Simple ETS forecast function
}
