% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{simulate from a forecasting function}
\usage{
simulator(
  FUN,
  y,
  h,
  level = 95,
  method = c("block-bootstrap", "surrogate", "kde", "bootstrap", "fitdistr", "meboot"),
  nsim = 100L,
  block_size = 5,
  seed = 123L,
  B = NULL,
  ...
)
}
\arguments{
\item{FUN}{A forecasting function.}

\item{y}{A time series (\code{ts} object or vector).}

\item{h}{Forecasting horizon.}

\item{level}{Confidence level.}

\item{method}{Method to be used for the simulation of in-sample residuals.}

\item{nsim}{Number of simulations.}

\item{block_size}{Block size for block-bootstrap.}

\item{seed}{Seed for reproducibility.}

\item{B}{Alias for \code{nsim}}

\item{...}{Additional arguments to be passed to the forecasting function.}
}
\value{
An object of class \code{forecast}, with a slot \code{sims}
}
\description{
This function allows to obtain predictive simulations from
any forecasting function.
}
\examples{



y <- fdeaths
h <- 25L
obj <- simulator(FUN=forecast::thetaf, y, h); plot(obj)
# obj <- simulator(FUN=forecast::auto.arima, y, h); plot(obj)
obj <- simulator(FUN=forecast::ets, y, h); plot(obj)
 
}
