% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthsimulation.R
\name{summary.synthetic_ts}
\alias{summary.synthetic_ts}
\title{Summary method for synthetic_ts objects}
\usage{
\method{summary}{synthetic_ts}(object, ...)
}
\arguments{
\item{object}{synthetic_ts object}

\item{...}{Additional arguments}
}
\description{
Summary method for synthetic_ts objects
}
