% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts2.R
\name{topdown_forecast}
\alias{topdown_forecast}
\title{Top-down forecast using historical proportions}
\usage{
topdown_forecast(hts, h)
}
\arguments{
\item{hts}{2-level HTS object}

\item{h}{Forecast horizon}
}
\value{
List with total and bottom forecasts
}
\description{
Top-down forecast using historical proportions
}
