% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transformerreturnsf}
\alias{transformerreturnsf}
\title{Transformer Forecasting for Financial Returns}
\usage{
transformerreturnsf(
  y,
  h = 10,
  level = 95,
  sequence_size = 20,
  num_heads = 4,
  head_size = 32,
  num_transformer_blocks = 2,
  epochs = 50,
  n_bootstrap = 250,
  show_progress = TRUE,
  seed = 123,
  ...
)
}
\arguments{
\item{y}{Univariate time series of returns}

\item{h}{Forecast horizon}

\item{level}{Confidence levels for prediction intervals}

\item{sequence_size}{Lookback period for transformer}

\item{...}{Additional parameters to TRANSFORMER function}
}
\value{
Object of class 'forecast'
}
\description{
Transformer Forecasting for Financial Returns
}
