% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnosticgarch.R
\name{agnosticgarchf}
\alias{agnosticgarchf}
\title{ANY MODEL+GARCH(1, 1) forecasting}
\usage{
agnosticgarchf(
  y,
  h = 5,
  level = 95,
  FUN = forecast::auto.arima,
  seed = 123,
  ...
)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{number of periods for forecasting}

\item{level}{confidence level for prediction intervals}

\item{FUN}{forecasting function for the main model;
default \code{ahead::dynrmf}}

\item{seed}{reproducibility seed}

\item{...}{}
}
\value{
An object of class "forecast"; a list containing the following elements:

\item{model}{A list containing information about the fitted model}
\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{x}{The original time series}
\item{sims}{Simulations of ANYMODEL+GARCH(1, 1)}
}
\description{
ANY MODEL+GARCH(1, 1) forecasting
}
\examples{

y <- datasets::EuStockMarkets[ , "DAX"]

# require(forecast)
# z <- ahead::agnosticgarchf(y=y, h=20)
# plot(z)

}
\author{
T. Moudiki
}
