% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basicf}
\alias{basicf}
\title{Basic forecasting (mean, median, random walk)}
\usage{
basicf(
  y,
  h = 5,
  level = 95,
  method = c("mean", "median", "rw"),
  type_pi = c("gaussian", "bootstrap"),
  seed = 1,
  B = 100
)
}
\arguments{
\item{y}{A multivariate time series of class \code{ts} or a matrix}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{method}{forecasting method, either "mean", "median", or random walk ("rw")}

\item{type_pi}{type of prediction interval currently, "gaussian" or "bootstrap"}

\item{seed}{reproducibility seed for \code{type_pi == 'bootstrap'}}

\item{B}{Number of bootstrap replications for \code{type_pi == 'bootstrap'}}
}
\description{
Basic forecasting functions for multivariate time series
}
\examples{

res <- ahead::basicf(fpp::insurance, h=10)
par(mfrow=c(1, 2))
plot(res, "TV.advert")
plot(res, "Quotes")


res <- ahead::basicf(fpp::insurance, method="rw", h=10)
par(mfrow=c(1, 2))
plot(res, "TV.advert")
plot(res, "Quotes")


}
