% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_Ridge.R
\name{comb_Ridge}
\alias{comb_Ridge}
\title{Ridge Regression Forecast Combination}
\usage{
comb_Ridge(x, custom_error = NULL)
}
\arguments{
\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}
}
\value{
Returns an object of class \code{ForecastComb::foreccomb_res} with the following components:
\item{Method}{Returns the best-fit forecast combination method.}
\item{Models}{Returns the individual input models that were used for the forecast combinations.}
\item{Weights}{Returns the combination weights obtained by applying the combination method to the training set.}
\item{Intercept}{Returns the intercept of the linear regression.}
\item{Fitted}{Returns the fitted values of the combination method for the training set.}
\item{Accuracy_Train}{Returns range of summary measures of the forecast accuracy for the training set.}
\item{Forecasts_Test}{Returns forecasts produced by the combination method for the test set. Only returned if input included a forecast matrix for the test set.}
\item{Accuracy_Test}{Returns range of summary measures of the forecast accuracy for the test set. Only returned if input included a forecast matrix and a vector of actual values for the test set.}
\item{Input_Data}{Returns the data forwarded to the method.}
}
\description{
Computes forecast combination weights using Ridge Regression (OLS) regression.
}
\details{
The function integrates the Ridge Regression forecast combination implementation of the
\emph{ForecastCombinations} package into ForecastComb.

The results are stored in an object of class 'ForecastComb::foreccomb_res', for which separate plot and summary functions are provided.
}
\examples{
obs <- rnorm(100)
preds <- matrix(rnorm(1000, 1), 100, 10)
train_o<-obs[1:80]
train_p<-preds[1:80,]
test_o<-obs[81:100]
test_p<-preds[81:100,]

data<-ForecastComb::foreccomb(train_o, train_p, test_o, test_p)
ahead::comb_Ridge(data)

}
\seealso{
\code{\link[ForecastCombinations]{Forecast_comb}},
\code{\link{foreccomb}},
\code{\link{plot.ForecastComb::foreccomb_res}},
\code{\link{summary.ForecastComb::foreccomb_res}},
\code{\link[forecast]{accuracy}}
}
\keyword{models}
