% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalize.R
\name{conformalize}
\alias{conformalize}
\title{Conformalize a forecasting function}
\usage{
conformalize(
  FUN,
  y,
  h,
  level = 95,
  method = c("block-bootstrap", "surrogate", "kde", "bootstrap", "fitdistr", "meboot"),
  nsim = 250L,
  block_size = 5,
  seed = 123L,
  B = NULL,
  ...
)
}
\arguments{
\item{FUN}{A forecasting function.}

\item{y}{A time series (\code{ts} object or vector).}

\item{h}{Forecasting horizon.}

\item{level}{Confidence level.}

\item{method}{Method to be used for conformalization (simulation of calibrated residuals).}

\item{nsim}{Number of simulations.}

\item{block_size}{Block size for block-bootstrap.}

\item{seed}{Seed for reproducibility.}

\item{B}{Alias for \code{nsim}}

\item{...}{Additional arguments to be passed to the forecasting function.}
}
\value{
An object of class \code{forecast}.
}
\description{
This function allows to conformalize any forecasting function.
}
\examples{

y <- fdeaths
h <- 25L
obj <- conformalize(FUN=forecast::ets, y, h); plot(obj)
obj <- conformalize(FUN=HoltWinters, y=y, h=h, seasonal = "mult"); plot(obj)

}
