% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eat.R
\name{eatf}
\alias{eatf}
\title{Combined ets-arima-theta forecasts}
\usage{
eatf(
  y,
  h = 5,
  level = 95,
  method = c("EAT", "E", "A", "T"),
  weights = rep(1/3, 3),
  type_pi = c("gaussian", "E", "A", "T"),
  ...
)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{number of periods for forecasting}

\item{level}{confidence level for prediction intervals}

\item{method}{forecasting method: "E" for \code{forecast::ets};
"A"for \code{forecast::auto.arima}; "T" for \code{forecast::thetaf};
or "EAT" for the combination of the three (default, with \code{weights})}

\item{weights}{weights for each method, in method \code{EAT}. Must add up to 1.}

\item{type_pi}{type of prediction interval:  currently ETS: "E", Auto.Arima: "A" or Theta: "T"}

\item{...}{additional parameters to be passed to \code{forecast::ets},
\code{forecast::auto.arima}, \code{forecast::thetaf} and
\code{forecast::forecast}}
}
\value{
An object of class "forecast"; a list containing the following elements:

\item{model}{A list containing information about the fitted model}
\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{x}{The original time series}
\item{residuals}{Residuals from the fitted model}
}
\description{
Combined ets, arima, and theta (eat) forecasting (uses \code{forecast::ets},
\code{forecast::auto.arima}, \code{forecast::thetaf})
}
\details{
ensemble forecasts obtained from \code{forecast::ets},
\code{forecast::auto.arima} and \code{forecast::theta} (with weights)
}
\examples{

require(forecast)

\dontrun{

print(ahead::eatf(WWWusage, method = "EAT",
weights = c(0.5, 0, 0.5)))

print(ahead::eatf(WWWusage, method = "EAT"))


obj <- ahead::eatf(WWWusage, method = "EAT",
weights = c(0, 0.5, 0.5), h=10,
type_pi = "T")
plot(obj)


obj <- ahead::eatf(WWWusage, method = "EAT",
weights = c(0, 0.5, 0.5), h=10, type_pi="A")
plot(obj)
}


par(mfrow=c(3, 2))
plot(ahead::eatf(USAccDeaths, h=10, level=95))
plot(ahead::eatf(AirPassengers, h=10, level=95, type_pi = "T"))
plot(ahead::eatf(lynx, h=10, level=95, type_pi = "A"))
plot(ahead::eatf(WWWusage, h=10, level=95, type_pi = "E"))
plot(ahead::eatf(Nile, h=10, level=95))
plot(ahead::eatf(fdeaths, h=10, level=95))


}
\references{
Hyndman R, Athanasopoulos G, Bergmeir C, Caceres G, Chhay L,
O'Hara-Wild M, Petropoulos F, Razbash S, Wang E, Yasmeen F (2021).
forecast: Forecasting functions for time series and linear models. R
package version 8.14, <URL: https://pkg.robjhyndman.com/forecast/>. \cr

Hyndman RJ, Khandakar Y (2008). 'Automatic time series forecasting: the
forecast package for R.' Journal of Statistical Software, 26 (3),
1-22. <URL: https://www.jstatsoft.org/article/view/v027i03>.

Assimakopoulos, V. and Nikolopoulos, K. (2000). The theta model: a
decomposition approach to forecasting. International Journal of
Forecasting 16, 521-530.

Hyndman, R.J., and Billah, B. (2003) Unmasking the Theta method.
International J. Forecasting, 19, 287-290.
}
\author{
T. Moudiki
}
