% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmtheta.R
\name{glmthetaf}
\alias{glmthetaf}
\title{Generalized Linear Model Theta Forecast}
\usage{
glmthetaf(
  y,
  h = ifelse(frequency(y) > 1, 2 * frequency(y), 10),
  level = 95L,
  fit_func = stats::glm,
  fan = FALSE,
  x = y,
  type_pi = c("conformal-split", "conformal-surrogate", "conformal-kde",
    "conformal-bootstrap", "conformal-block-bootstrap", "conformal-fitdistr", "gaussian"),
  attention = TRUE,
  scale_ctxt = 1,
  B = 250L,
  nsim = B,
  ...
)
}
\arguments{
\item{y}{The time series data}

\item{h}{The number of periods to forecast}

\item{level}{The confidence level for the forecast intervals}

\item{fit_func}{The function to use for fitting the GLM}

\item{fan}{Logical flag for fan plot}

\item{x}{The time series data}

\item{attention}{Logical flag for using attention mechanism}

\item{scale_ctxt}{Scaling coefficient for context vector}

\item{B}{Number of bootstrap replications or number of simulations (yes, 'B' is unfortunate)}

\item{nsim}{Alias for B}

\item{...}{Additional arguments to pass to the fit_func}
}
\value{
A forecast object
}
\description{
This function implements the Theta method using a Generalized Linear Model (GLM)
}
