% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmtheta.R
\name{glmthetaf}
\alias{glmthetaf}
\title{Generalized Linear Model Theta Forecast}
\usage{
glmthetaf(
  y,
  h = ifelse(frequency(y) > 1, 2 * frequency(y), 10),
  level = c(80, 95),
  fit_func = lsfit,
  fan = FALSE,
  x = y,
  ...
)
}
\arguments{
\item{y}{The time series data}

\item{h}{The number of periods to forecast}

\item{level}{The confidence level for the forecast intervals}

\item{fit_func}{The function to use for fitting the GLM}

\item{fan}{Logical flag for fan plot}

\item{x}{The time series data}

\item{...}{Additional arguments to pass to the fit_func}
}
\value{
A forecast object
}
\description{
This function implements the Theta method using a Generalized Linear Model (GLM)
}
