% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{loessf}
\alias{loessf}
\title{Loess forecasting}
\usage{
loessf(
  y,
  h = 5,
  level = 95,
  span = 0.75,
  degree = 2,
  b = NULL,
  B = 250,
  seed = 123
)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{span}{the parameter which controls the degree of smoothing}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2. (Degree 0 is also allowed, but see \code{stats::loess})}

\item{b}{block length for circular block bootstrap}

\item{B}{number of bootstrap replications}

\item{seed}{reproducibility seed}
}
\description{
Loess forecasting
}
\examples{

plot(loessf(Nile, h=20, level=95))

}
