% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{loessf}
\alias{loessf}
\title{Loess forecasting}
\usage{
loessf(
  y,
  h = 5,
  level = 95,
  span = 0.75,
  degree = 2,
  type_pi = c("bootstrap", "blockbootstrap", "movingblockbootstrap"),
  b = NULL,
  B = 250,
  type_aggregation = c("mean", "median"),
  seed = 123
)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{span}{the parameter which controls the degree of smoothing}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2. (Degree 0 is also allowed, but see \code{stats::loess})}

\item{type_pi}{Type of prediction interval currently (independent) "bootstrap", (circular) "blockbootstrap", or "movingblockbootstrap"}

\item{b}{block length for circular block bootstrap}

\item{B}{number of bootstrap replications}

\item{type_aggregation}{Type of aggregation, ONLY for bootstrapping; either "mean" or "median"}

\item{seed}{reproducibility seed}
}
\value{
An object of class "forecast"; a list containing the following elements:

\item{model}{A list containing information about the fitted model}
\item{method}{The name of the forecasting method as a character string}
\item{mean}{Point forecasts for the time series}
\item{lower}{Lower bound for prediction interval}
\item{upper}{Upper bound for prediction interval}
\item{x}{The original time series}
\item{residuals}{Residuals from the fitted model}
\item{sims}{Model simulations for bootstrapping}
}
\description{
Loess forecasting
}
\examples{

par(mfrow = c(3, 1))

plot(loessf(Nile, h=20, level=95, B=10))

plot(loessf(Nile, h=20, level=95, B=10,
     type_pi = "blockbootstrap"))

plot(loessf(Nile, h=20, level=95, B=10,
     type_pi = "movingblockbootstrap"))

}
