% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{mlf}
\alias{mlf}
\title{Conformalized Forecasting using Machine Leaning models}
\usage{
mlf(
  y,
  h = 5,
  level = 95,
  lags = 15L,
  fit_func = ahead::ridge,
  predict_func = predict,
  type_pi = c("kde", "surrogate", "blockbootstrap"),
  B = 250L,
  agg = c("mean", "median"),
  seed = 123,
  ...
)
}
\arguments{
\item{y}{A numeric vector or time series of class \code{ts}}

\item{h}{Forecasting horizon}

\item{level}{Confidence level for prediction intervals}

\item{lags}{Number of lags of the input time series considered in the regression}

\item{fit_func}{Fitting function (Statistical/ML model). Default is Ridge regression.}

\item{predict_func}{Prediction function (Statistical/ML model)}

\item{type_pi}{Type of prediction interval}

\item{B}{Number of bootstrap replications or number of simulations}

\item{agg}{"mean" or "median" (aggregation method)}

\item{...}{additional parameters passed to the fitting function \code{fit_func}}
}
\description{
Conformalized Forecasting using Machine Leaning models
}
\examples{

res <- ahead::mlf(USAccDeaths, h=10L, lags=15L, type_pi="surrogate", B=250L)
plot(res)

res <- ahead::mlf(USAccDeaths, fit_func = glmnet::cv.glmnet, h=15L, lags=15L, 
type_pi="kde", B=250L) 
plot(res)

(res <- ahead::mlf(USAccDeaths, fit_func = e1071::svm, h=15L, lags=15L, 
type_pi="kde", B=250L)) 
plot(res)

}
