% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts2.R
\name{plot_hts_forecast}
\alias{plot_hts_forecast}
\title{Plot forecast results with simulation intervals}
\usage{
plot_hts_forecast(result, hts, series_type = "total", series_idx = 1)
}
\description{
Plot forecast results with simulation intervals
}
