% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hts2.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Plot simulation paths}
\usage{
plot_simulations(
  result,
  series_type = "total",
  series_idx = 1,
  n_paths = 50,
  main = NULL
)
}
\description{
Plot simulation paths
}
