% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretfitpredict.R
\name{predict_func}
\alias{predict_func}
\title{Predict univariate time series using caret ML model(for use with \code{dynrmf})}
\usage{
predict_func(obj, newx)
}
\arguments{
\item{x}{A matrix of predictors}

\item{y}{A vector of responses}

\item{method}{The caret method to use for fitting the model}

\item{verbose}{Whether to print the model summary}
}
\value{
A model object
}
\description{
Predict univariate time series using caret ML model(for use with \code{dynrmf})
}
