% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Fit Ridge regression}
\usage{
ridge(x, y, lambda = 10^seq(-10, 10, length.out = 100))
}
\arguments{
\item{x}{matrix of examples}

\item{y}{a vector, the response}

\item{lambda}{regularization parameters}
}
\value{
a list, an object of class 'ridge'
}
\description{
Fit Ridge regression
}
\examples{

n <- 100 ; p <- 10
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

(fit_obj <- ahead::ridge(X, y))

matplot(log(fit_obj$lambda), t(fit_obj$coef), type = 'l')

}
