% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitts.R
\name{splitts}
\alias{splitts}
\title{Partition a time series object}
\usage{
splitts(y, split_prob = 0.5, return_indices = FALSE, ...)
}
\arguments{
\item{y}{A time series object}

\item{split_prob}{Splitting ratio}

\item{return_indices}{if TRUE, returns series' indices, otherwise, time series objects}
}
\description{
Partition a time series object
}
\examples{

misc::splitts(ts(1:10))

}
