% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.foreccomb_res.R
\name{summary.foreccomb_res}
\alias{summary.foreccomb_res}
\alias{print.foreccomb_res_summary}
\title{Summary of Forecast Combination}
\usage{
\method{summary}{foreccomb_res}(object, ...)

\method{print}{foreccomb_res_summary}(x, ...)
}
\arguments{
\item{object}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}

\item{...}{potential further arguments (require by generic)}

\item{x}{An object of class 'foreccomb'. Contains training set (actual values + matrix of model forecasts) and optionally a test set.}
}
\description{
\code{summary} method for class \sQuote{foreccomb_res}. Includes information about combination method,
combination weights assigned to the individual forecast models, as well as an accuracy evaluation of the combined
forecast.
}
\seealso{
\code{\link[ForecastComb]{foreccomb}},
\code{\link[ForecastComb]{plot.foreccomb_res}},
}
\author{
Christoph E. Weiss and Gernot R. Roetzer
}
